+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.1
+- Added Json Formatter
+- Updated Example and readme
+- ColoredLog based on value type
+
+
 ## 1.0.0
 
 - Initial release of the ColoredLog package with rich logging features
@@ -6,3 +12,4 @@
 - Support for text stylization (bold, italic, underline, etc.).
 - Option to make logs blink.
 - Simple and intuitive API for logging with customization options.
+